refactor: split capacity-limit test files, duplicate collapse, and ui shards
PR additions landed on surfaces parked exactly at their caps. Split the
forwarded/tool-details projection suites and sender-label/sender-session
normalizer suites out of their 1000-line test files, extract the
sequential-duplicate-collapse concept from chat-thread-items (706>700)
into chat-thread-duplicates, and give ui/src/e2e its own tsgo shard so
ui-pages-e2e returns under the 720-root memory bound. Pure moves; no
caps or baselines changed.
